Added explicit color conversion

+++ projects/mouse_burrows/algorithm/processes/predug_detector.py
@@ -54,7 +54,7 @@ class PredugDetector(object):
             """ fills the mask with the buffered regions """
             for poly in (poly_p, poly_s):
                 pts = np.array(poly.buffer(margin).boundary.coords, np.int32)
-                cv2.fillPoly(mask, [pts], color)
+                cv2.fillPoly(mask, [pts], np.uint8(color))
                 
         # prepare the mask for the grabCut algorithm
         burrow_width = self.params['burrows/width']
